refactor: divide api service & retrofit module
